// src/app/interfaces/packages.ts
export interface Package {
  fullAdress: string;
  coordinates: string;
  id: number;
  clientname: string;
  image: string;
  quantity: number;
  weight: number;
  deliveryday: Date;
  street: string;
  number: number;
  city: string;
  province: string;
  postalCode: string;
  lat: number;
  lng: number;
  status: string;
}